+++ demo/docs/css/app.1ed23334.css
@@ -1,0 +1,1 @@
+#app{font-family:Avenir,Helvetica,Arial,sans-serif;-webkit-font-smoothing:antialiased;-moz-osx-font-smoothing:grayscale;color:#2c3e50}#app .container{max-width:700px}.nav{padding:30px}.nav a{font-weight:700;color:#2c3e50}.nav a.router-link-exact-active{color:#42b983}.details[data-v-21083be0]{border:1px solid gray;border-radius:4px;padding:.5em .5em 0}.details[data-open][data-v-21083be0]{padding:.5em}.summary[data-v-21083be0]{font-weight:700;margin:-.5em -.5em 0;padding:.5em;cursor:pointer}.summary .summary-icon-opened[data-v-21083be0]{display:inline}.details-body[data-v-21083be0],.summary .summary-icon-closeed[data-v-21083be0]{display:none}.details-body[data-open][data-v-21083be0]{display:block}.summary .summary-opened[data-v-00ab0eee]{display:none}.summary .summary-closed[data-v-00ab0eee],.summary[data-open] .summary-opened[data-v-00ab0eee]{display:block}.summary[data-open] .summary-closed[data-v-00ab0eee]{display:none}
