+++ demo/docs/js/about.d8398a3f.js
@@ -1,0 +1,2 @@
+"use strict";(self["webpackChunkdemo"]=self["webpackChunkdemo"]||[]).push([[443],{3008:function(i,s,t){t.r(s),t.d(s,{default:function(){return b}});var o=t(3396);const a={class:"about"};function r(i,s,t,r,l,u){const e=(0,o.up)("PAboutVue");return(0,o.wg)(),(0,o.iD)("div",a,[(0,o.Wm)(e)])}t(8010);const l={class:"container"},u=(0,o.uE)('<div class="list-group list-group-flush"><div class="list-group-item">Name</div><div class="list-group-item"><div class="list-group list-group-flush"><div class="list-group-item"><code>@rksan/random-string</code></div></div></div><div class="list-group-item">About</div><div class="list-group-item"><div class="list-group list-group-flush"><div class="list-group-item"> This application is a demonstration page of a module that generates random strings. </div></div></div><div class="list-group-item">Licence</div><div class="list-group-item"><div class="list-group list-group-flush"><div class="list-group-item">MIT</div></div></div><div class="list-group-item"></div><div class="list-group-item">Links</div><div class="list-group-item"><div class="list-group list-group-flush"><a class="list-group-item list-group-item-action" href="https://github.com/rksan/random-string" target="_blank"><i class="bi bi-github"></i><code> Github </code><i class="bi bi-box-arrow-up-right"></i></a><a class="list-group-item list-group-item-action" href="https://www.npmjs.com/package/@rksan/random-string" target="_blank"><code> npm </code><i class="bi bi-box-arrow-up-right"></i></a></div></div></div>',1),e=[u];function d(i,s){return(0,o.wg)(),(0,o.iD)("div",l,e)}var c=t(89);const n={},g=(0,c.Z)(n,[["render",d]]);var p=g,v=(0,o.aZ)({name:"about-view",components:{PAboutVue:p}});const m=(0,c.Z)(v,[["render",r]]);var b=m}}]);
+//# sourceMappingURL=about.d8398a3f.js.map
