+++ demo/docs/js/app.8eaa8216.js
@@ -1,0 +1,2 @@
+(function(){"use strict";var e={8010:function(e,t,n){n.d(t,{Z:function(){return tt}});var o=n(3396),a=n(7139),r=n(9242);const s=e=>((0,o.dD)("data-v-b3d70384"),e=e(),(0,o.Cn)(),e),i={class:"container"},l=s((()=>(0,o._)("h1",null,"Random String Demo",-1))),d={class:"form-floating mb-3"},c=["placeholder"],u=s((()=>(0,o._)("label",{for:"result",class:"form-label"},(0,a.zw)("Result"),-1))),m={class:"row gy-1 gx-2 align-items-center"},g={class:"col-auto"},p=s((()=>(0,o._)("div",{class:"col-2"},[(0,o._)("button",{class:"btn btn-warning"},"Clear")],-1))),f={class:"raw mb-3"},v=s((()=>(0,o._)("legend",{class:"col form-label"},"aruguments",-1))),b={key:0,class:"row mb-3"},h={class:"col"},y={class:"lead bg-danger text-light"},_={class:"row mb-3 align-items-center"},w={class:"col-auto"},k={class:"form-floating"},C=s((()=>(0,o._)("label",{for:"input-length-number",class:"form-label"},"Length",-1))),x={class:"col"},O=(0,o.uE)('<datalist id="input-length-range-list" data-v-b3d70384><option value="1" data-v-b3d70384>1</option><option value="8" data-v-b3d70384>8</option><option value="16" data-v-b3d70384>16</option><option value="32" data-v-b3d70384>32</option><option value="64" data-v-b3d70384>64</option><option value="128" data-v-b3d70384>128</option></datalist>',1),D={class:"row"},S={class:"col"},j={class:"row"},A={class:"col"};function V(e,t){const n=(0,o.up)("OOptionsVue");return(0,o.wg)(),(0,o.iD)("div",i,[l,(0,o._)("form",{novalidate:"",class:(0,a.C_)(e.classes.form)},[(0,o._)("div",d,[(0,o.wy)((0,o._)("textarea",{id:"result",class:"form-control",name:"Result","onUpdate:modelValue":t[0]||(t[0]=t=>e.data.result=t),placeholder:`${e.data.sample}`},null,8,c),[[r.nr,e.data.result]]),u]),(0,o._)("div",m,[(0,o._)("div",g,[(0,o._)("button",{class:"btn btn-primary",onClick:t[1]||(t[1]=(...t)=>e.doClickRandomString&&e.doClickRandomString(...t))}," Generate Random String ")]),p]),(0,o._)("fieldset",f,[v,e.data.message?((0,o.wg)(),(0,o.iD)("div",b,[(0,o._)("div",h,[(0,o._)("p",y,(0,a.zw)(`invalid : ${e.data.message}`),1)])])):(0,o.kq)("",!0),(0,o._)("div",_,[(0,o._)("div",w,[(0,o._)("div",k,[(0,o.wy)((0,o._)("input",{type:"number",id:"input-length-number",class:"form-control",min:"1",max:"128",list:"input-length-range-list","onUpdate:modelValue":t[2]||(t[2]=t=>e.model.length=t),required:""},null,512),[[r.nr,e.model.length]]),C])]),(0,o._)("div",x,[(0,o.wy)((0,o._)("input",{type:"range",id:"input-length-range",style:{width:"100%"},min:"1",max:"128",list:"input-length-range-list","onUpdate:modelValue":t[3]||(t[3]=t=>e.model.length=t)},null,512),[[r.nr,e.model.length]])]),O]),(0,o._)("div",D,[(0,o._)("div",S,[(0,o.Wm)(n,{id:"options-src",title:"src",args:e.model.src,onChangeOptions:e.fireChangeOptionsSrc,reqired:""},null,8,["args","onChangeOptions"])])]),(0,o._)("div",j,[(0,o._)("div",A,[(0,o.Wm)(n,{id:"options-exclude",title:"exclude",args:e.model.exclude,onChangeOptions:e.fireChangeOptionsExclude},null,8,["args","onChangeOptions"])])])])],2)])}n(7658);var E=n(4870);const Z=e=>((0,o.dD)("data-v-5b3bb932"),e=e(),(0,o.Cn)(),e),$=["id"],q={class:"col-auto"},U={class:"form-floating"},P=["id"],T=Z((()=>(0,o._)("option",{value:"text"},"text",-1))),H=Z((()=>(0,o._)("option",{value:"range"},"range",-1))),I=[T,H],K=["for"],M={class:"col"},W={class:"form-floating"},z=["id"],L=["value"],R=["for"],Y={class:"col"},F={class:"form-floating"},N=["id"],B=["value"],G=["for"],J={key:1,class:"col"},Q={class:"form-floating"},X=["id"],ee=["for"];function te(e,t){return(0,o.wg)(),(0,o.iD)("div",{class:"row",id:e.id},[(0,o._)("div",q,[(0,o._)("div",U,[(0,o.wy)((0,o._)("select",{class:"form-select arg-kind",id:`kind-${e.id}`,"onUpdate:modelValue":t[0]||(t[0]=t=>e.model.kind=t),onChange:t[1]||(t[1]=(...t)=>e.doChangeKind&&e.doChangeKind(...t))},I,40,P),[[r.bM,e.model.kind]]),(0,o._)("label",{for:`kind-${e.id}`},"Kind",8,K)])]),"range"===e.model.kind?((0,o.wg)(),(0,o.iD)(o.HY,{key:0},[(0,o._)("div",M,[(0,o._)("div",W,[(0,o.wy)((0,o._)("select",{id:`start-${e.id}`,class:"form-select arg-start","onUpdate:modelValue":t[2]||(t[2]=t=>e.model.start=t),onChange:t[3]||(t[3]=(...t)=>e.doChangeStart&&e.doChangeStart(...t))},[((0,o.wg)(!0),(0,o.iD)(o.HY,null,(0,o.Ko)(e.model.options,((e,t)=>((0,o.wg)(),(0,o.iD)("option",{key:t,value:e},(0,a.zw)(e),9,L)))),128))],40,z),[[r.bM,e.model.start]]),(0,o._)("label",{for:`start-${e.id}`,class:"form-label"},"Start",8,R)])]),(0,o._)("div",Y,[(0,o._)("div",F,[(0,o.wy)((0,o._)("select",{id:`end-${e.id}`,class:"form-select","onUpdate:modelValue":t[4]||(t[4]=t=>e.model.end=t),onChange:t[5]||(t[5]=(...t)=>e.doChangeEnd&&e.doChangeEnd(...t))},[((0,o.wg)(!0),(0,o.iD)(o.HY,null,(0,o.Ko)(e.model.options,((e,t)=>((0,o.wg)(),(0,o.iD)("option",{key:t,value:e},(0,a.zw)(e),9,B)))),128))],40,N),[[r.bM,e.model.end]]),(0,o._)("label",{for:`end-${e.id}`,class:"from-label"},"End",8,G)])])],64)):((0,o.wg)(),(0,o.iD)("div",J,[(0,o._)("div",Q,[(0,o.wy)((0,o._)("textarea",{id:`text-${e.id}`,class:"form-control","onUpdate:modelValue":t[6]||(t[6]=t=>e.model.text=t)},null,8,X),[[r.nr,e.model.text]]),(0,o._)("label",{for:`text-${e.id}`,class:"from-label"},"Text",8,ee)])]))],8,$)}const ne=(e,t)=>{const n=[],o="string"===typeof e?e.charCodeAt(0):e,a="string"===typeof t?t.charCodeAt(0):t;for(let r=o;r<=a;r++)n.push(String.fromCharCode(r));return n},oe={id:{type:String},kind:{type:String,default:()=>"range"},arg:{type:[String,Object],defalt:()=>({start:"!",end:"~"})},options:{type:Array,default:()=>{}},require:{type:Boolean,default:()=>!1}},ae=["change-options-arg","change-options-arg-kind","change-options-arg-start","change-options-arg-end"],re=(e,t)=>{const n=t.attrs["data-list-index"],o=e.kind,a=n?parseInt(n,10):0,r=(()=>{let t="";return"object"===typeof e.arg&&(t=e.arg.start),t||"!"})(),s=(()=>{let t="";return"object"===typeof e.arg&&(t=e.arg.end),t||"~"})(),i=(()=>{let t="";return"string"===typeof e.arg&&(t=e.arg),t||"0123456789ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyx"})(),l=e.options||ne(r,s),d=(0,E.qj)({index:a,kind:o,start:r,end:s,text:i,options:l}),c=()=>{const e={index:d.index,kind:d.kind,arg:{}};return"range"===d.kind?e.arg={start:d.start,end:d.end}:e.arg={text:d.text},e},u=e=>{e&&e.preventDefault();const n=c();t.emit("change-options-arg-kind",{model:n}),t.emit("change-options-arg",{model:n})},m=e=>{e&&e.preventDefault();const n=c();t.emit("change-options-arg-start",{model:n}),t.emit("change-options-arg",{model:n})},g=e=>{e&&e.preventDefault();const n=c();t.emit("change-options-arg-end",{model:n}),t.emit("change-options-arg",{model:n})};return{model:d,doChangeKind:u,doChangeStart:m,doChangeEnd:g}};var se=(0,o.aZ)({name:"o-options-arg",props:oe,emits:ae,setup:re}),ie=n(89);const le=(0,ie.Z)(se,[["render",te],["__scopeId","data-v-5b3bb932"]]);var de=le,ce=de;const ue=e=>((0,o.dD)("data-v-00ab0eee"),e=e(),(0,o.Cn)(),e),me={class:"summary-closed"},ge={class:"row"},pe={class:"col"},fe=ue((()=>(0,o._)("i",{class:"bi bi-caret-right summary-icon-closed"},null,-1))),ve={class:"summary-opened"},be={class:"row"},he={class:"col"},ye=ue((()=>(0,o._)("i",{class:"bi bi-caret-down summary-icon-opened"},null,-1))),_e={class:"col-auto"},we=ue((()=>(0,o._)("i",{class:"bi bi-plus-circle"},null,-1))),ke=[we],Ce={class:"col"},xe={class:"col-auto"},Oe=["data-list-index"],De=ue((()=>(0,o._)("i",{class:"bi bi-dash-circle"},null,-1))),Se=[De];function je(e,t){const n=(0,o.up)("OOptionsArgVue"),r=(0,o.up)("MDetailsVue");return(0,o.wg)(),(0,o.j4)(r,{summary:e.title,id:e.id,class:"mb-3"},{summary:(0,o.w5)((()=>[(0,o._)("div",me,[(0,o._)("div",ge,[(0,o._)("div",pe,[fe,(0,o.Uk)(" "+(0,a.zw)(e.title),1)])])]),(0,o._)("div",ve,[(0,o._)("div",be,[(0,o._)("div",he,[ye,(0,o.Uk)(" "+(0,a.zw)(e.title),1)]),(0,o._)("div",_e,[(0,o._)("button",{class:"btn btn-success",onClick:t[0]||(t[0]=(...t)=>e.doClickAdd&&e.doClickAdd(...t)),style:{"z-index":"1"}},ke)])])])])),default:(0,o.w5)((()=>[((0,o.wg)(!0),(0,o.iD)(o.HY,null,(0,o.Ko)(e.model.args,((a,r)=>((0,o.wg)(),(0,o.iD)("div",{key:r,class:"row mb-3"},[(0,o._)("div",Ce,[(0,o.Wm)(n,{id:`${e.id}-${r}`,"data-list-index":r,options:e.options,arg:a,onChangeOptionsArg:e.fireChangeOptions},null,8,["id","data-list-index","options","arg","onChangeOptionsArg"])]),(0,o._)("div",xe,[(0,o._)("button",{class:"btn btn-danger",onClick:t[1]||(t[1]=(...t)=>e.doClickDel&&e.doClickDel(...t)),"data-list-index":r},Se,8,Oe)])])))),128))])),_:1},8,["summary","id"])}const Ae=e=>((0,o.dD)("data-v-21083be0"),e=e(),(0,o.Cn)(),e),Ve={class:"details"},Ee={key:0,class:"summary-label"},Ze=Ae((()=>(0,o._)("i",{class:"bi bi-caret-right summary-icon-closed"},null,-1))),$e=Ae((()=>(0,o._)("i",{class:"bi bi-caret-down summary-icon-opened"},null,-1))),qe={class:"details-body"};function Ue(e,t){return(0,o.wg)(),(0,o.iD)("div",Ve,[(0,o._)("div",{class:"summary",onClick:t[0]||(t[0]=(...t)=>e.doClickSummary&&e.doClickSummary(...t))},[(0,o.WI)(e.$slots,"summary",{},(()=>[e.summary?((0,o.wg)(),(0,o.iD)("div",Ee,[Ze,$e,(0,o.Uk)(" "+(0,a.zw)(e.summary),1)])):(0,o.kq)("",!0)]),!0)]),(0,o._)("div",qe,[(0,o.WI)(e.$slots,"default",{},void 0,!0)])])}const Pe={summary:{type:String}},Te=(e,t)=>{const n=e.parentElement;if(!n)return n;let o=!1;return t&&(!o&&t.tag&&(o=n.tagName.toLowerCase()===t.tag.toLowerCase()),!o&&t.cls&&(o=n.classList.contains(t.cls))),o?n:Te(n,t)},He=()=>{const e=e=>{e&&(e.preventDefault(),e.stopPropagation());const t=e.currentTarget;if(!t)return;const n=t,o=Te(n,{cls:"details"});if(!o)return;const a=o.querySelector(".details-body");n&&a&&(""===o.dataset.open?(o.removeAttribute("data-open"),n.removeAttribute("data-open"),a.removeAttribute("data-open")):(o.dataset.open="",n.dataset.open="",a.dataset.open=""))};return{doClickSummary:e}};var Ie=(0,o.aZ)({name:"m-details",props:Pe,setup:He});const Ke=(0,ie.Z)(Ie,[["render",Ue],["__scopeId","data-v-21083be0"]]);var Me=Ke,We=Me;const ze={id:{type:String},title:{type:String,default:()=>"arg-name"},args:{type:[Array,void 0],default:()=>{}}},Le=["change-options"],Re=(e,t)=>{const n=e=>{const t=new Array(e);for(var n=0;n<e;n++)t[n]={start:"!",end:"~"};return t},a=(0,E.qj)({length:e.length,args:e.args||n(1)}),r=(0,o.Fl)((()=>ne("!","~"))),s=e=>{const t=e.model.index,n=e.model.kind,o=e.model.arg;if("range"===n){const e={start:o.start,end:o.end};a.args[t]=e}else{const e=o.text;a.args[t]=e}},i=e=>{console.log("data",e),s(e),t.emit("change-options",{model:a})},l=e=>{e&&(e.preventDefault(),e.stopPropagation()),a.args.push({start:"!",end:"~"}),t.emit("change-options",{model:a})},d=e=>{e&&e.preventDefault();const n=e.currentTarget;if(!n)return;const o=n,r=o.dataset.listIndex;if(void 0===r)return;const s=parseInt(r,10);a.args=a.args.filter(((e,t)=>t!==s)),a.length=a.args.length,t.emit("change-options",{model:a})};return{model:a,options:r,fireChangeOptions:i,doClickAdd:l,doClickDel:d}};var Ye=(0,o.aZ)({name:"o-options",components:{OOptionsArgVue:ce,MDetailsVue:We},props:ze,emits:Le,setup:Re});const Fe=(0,ie.Z)(Ye,[["render",je],["__scopeId","data-v-00ab0eee"]]);var Ne=Fe,Be=Ne,Ge=n(9914);const Je=()=>{const e=(0,E.qj)({form:""}),t=(0,E.qj)({sample:(0,Ge.randomString)(8),result:"",message:""}),n=(0,E.qj)({length:8,src:[],exclude:[]});n.src.push({start:"!",end:"~"});const a=(0,o.Fl)((()=>ne("!","~"))),r=e=>{const t=e.model.args;n.src=t,l()},s=e=>{const t=e.model.args;n.exclude=t,l()},i=e=>{for(let t=0;t<e.length;t++)return e[t];return"object"},l=()=>{e.form="",t.message=""},d=()=>n.length<=0?{name:"length",message:"Must be 0 < Length"}:n.src.length<1?{name:"src",message:"One or more Src must be specified"}:n.src.every(((e,t,n)=>typeof e===typeof n[0]))?0<n.exclude.length&&!n.exclude.every(((e,t,n)=>typeof e===typeof n[0]))?{name:"exclude",message:"Kind of Exclude must be one type"}:void 0:{name:"src",message:"Kind of Src must be one type"},c=o=>{o&&o.preventDefault();let a=d();if(a)return e.form+="",void(t.message=a.message);const r=n.length,s="object"===i(n.src)?n.src.filter(((e,t,n)=>typeof e===typeof n[0])):n.src[0],l="object"===i(n.exclude)?n.exclude.filter(((e,t,n)=>typeof e===typeof n[0])):n.exclude[0];console.log("model.src=",n.src,"model.exclude=",n.exclude),console.log("src=",s,"exclude=",l);const c=(0,Ge.randomString)(r,{src:s,exclude:l});t.result=c},u=e=>{e&&e.preventDefault(),t.result=""};return{classes:e,data:t,model:n,options:a,fireChangeOptionsSrc:r,fireChangeOptionsExclude:s,doClickRandomString:c,doClickClear:u}};var Qe=(0,o.aZ)({name:"page-home",components:{OOptionsArgVue:ce,OOptionsVue:Be},setup:Je});const Xe=(0,ie.Z)(Qe,[["render",V],["__scopeId","data-v-b3d70384"]]);var et=Xe,tt=et},1580:function(e,t,n){var o=n(9242),a=n(3396);const r={class:"nav nav-underline justify-content-center"};function s(e,t){const n=(0,a.up)("router-link"),o=(0,a.up)("router-view");return(0,a.wg)(),(0,a.iD)(a.HY,null,[(0,a._)("nav",r,[(0,a.Wm)(n,{to:"/",class:"nav-link","aria-current":"page"},{default:(0,a.w5)((()=>[(0,a.Uk)("Home")])),_:1}),(0,a.Wm)(n,{to:"/about",class:"nav-link"},{default:(0,a.w5)((()=>[(0,a.Uk)("About")])),_:1})]),(0,a.Wm)(o)],64)}var i=n(89);const l={},d=(0,i.Z)(l,[["render",s]]);var c=d,u=n(2483);const m={class:"home"};function g(e,t,n,o,r,s){const i=(0,a.up)("PHomeVue");return(0,a.wg)(),(0,a.iD)("div",m,[(0,a.Wm)(i)])}var p=n(8010),f=(0,a.aZ)({name:"HomeView",components:{PHomeVue:p.Z}});const v=(0,i.Z)(f,[["render",g]]);var b=v;const h=[{path:"/",name:"home",component:b},{path:"/about",name:"about",component:()=>n.e(443).then(n.bind(n,3008))}],y=(()=>"/rksan/random-string")(),_=(0,u.p7)({history:(0,u.r5)(y),routes:h});var w=_;(0,o.ri)(c).use(w).mount("#app")}},t={};function n(o){var a=t[o];if(void 0!==a)return a.exports;var r=t[o]={exports:{}};return e[o].call(r.exports,r,r.exports,n),r.exports}n.m=e,function(){var e=[];n.O=function(t,o,a,r){if(!o){var s=1/0;for(c=0;c<e.length;c++){o=e[c][0],a=e[c][1],r=e[c][2];for(var i=!0,l=0;l<o.length;l++)(!1&r||s>=r)&&Object.keys(n.O).every((function(e){return n.O[e](o[l])}))?o.splice(l--,1):(i=!1,r<s&&(s=r));if(i){e.splice(c--,1);var d=a();void 0!==d&&(t=d)}}return t}r=r||0;for(var c=e.length;c>0&&e[c-1][2]>r;c--)e[c]=e[c-1];e[c]=[o,a,r]}}(),function(){n.n=function(e){var t=e&&e.__esModule?function(){return e["default"]}:function(){return e};return n.d(t,{a:t}),t}}(),function(){n.d=function(e,t){for(var o in t)n.o(t,o)&&!n.o(e,o)&&Object.defineProperty(e,o,{enumerable:!0,get:t[o]})}}(),function(){n.f={},n.e=function(e){return Promise.all(Object.keys(n.f).reduce((function(t,o){return n.f[o](e,t),t}),[]))}}(),function(){n.u=function(e){return"js/about.d8398a3f.js"}}(),function(){n.miniCssF=function(e){}}(),function(){n.g=function(){if("object"===typeof globalThis)return globalThis;try{return this||new Function("return this")()}catch(e){if("object"===typeof window)return window}}()}(),function(){n.o=function(e,t){return Object.prototype.hasOwnProperty.call(e,t)}}(),function(){var e={},t="demo:";n.l=function(o,a,r,s){if(e[o])e[o].push(a);else{var i,l;if(void 0!==r)for(var d=document.getElementsByTagName("script"),c=0;c<d.length;c++){var u=d[c];if(u.getAttribute("src")==o||u.getAttribute("data-webpack")==t+r){i=u;break}}i||(l=!0,i=document.createElement("script"),i.charset="utf-8",i.timeout=120,n.nc&&i.setAttribute("nonce",n.nc),i.setAttribute("data-webpack",t+r),i.src=o),e[o]=[a];var m=function(t,n){i.onerror=i.onload=null,clearTimeout(g);var a=e[o];if(delete e[o],i.parentNode&&i.parentNode.removeChild(i),a&&a.forEach((function(e){return e(n)})),t)return t(n)},g=setTimeout(m.bind(null,void 0,{type:"timeout",target:i}),12e4);i.onerror=m.bind(null,i.onerror),i.onload=m.bind(null,i.onload),l&&document.head.appendChild(i)}}}(),function(){n.r=function(e){"undefined"!==typeof Symbol&&Symbol.toStringTag&&Object.defineProperty(e,Symbol.toStringTag,{value:"Module"}),Object.defineProperty(e,"__esModule",{value:!0})}}(),function(){n.p=""}(),function(){var e={143:0};n.f.j=function(t,o){var a=n.o(e,t)?e[t]:void 0;if(0!==a)if(a)o.push(a[2]);else{var r=new Promise((function(n,o){a=e[t]=[n,o]}));o.push(a[2]=r);var s=n.p+n.u(t),i=new Error,l=function(o){if(n.o(e,t)&&(a=e[t],0!==a&&(e[t]=void 0),a)){var r=o&&("load"===o.type?"missing":o.type),s=o&&o.target&&o.target.src;i.message="Loading chunk "+t+" failed.\n("+r+": "+s+")",i.name="ChunkLoadError",i.type=r,i.request=s,a[1](i)}};n.l(s,l,"chunk-"+t,t)}},n.O.j=function(t){return 0===e[t]};var t=function(t,o){var a,r,s=o[0],i=o[1],l=o[2],d=0;if(s.some((function(t){return 0!==e[t]}))){for(a in i)n.o(i,a)&&(n.m[a]=i[a]);if(l)var c=l(n)}for(t&&t(o);d<s.length;d++)r=s[d],n.o(e,r)&&e[r]&&e[r][0](),e[r]=0;return n.O(c)},o=self["webpackChunkdemo"]=self["webpackChunkdemo"]||[];o.forEach(t.bind(null,0)),o.push=t.bind(null,o.push.bind(o))}();var o=n.O(void 0,[998],(function(){return n(1580)}));o=n.O(o)})();
+//# sourceMappingURL=app.8eaa8216.js.map
