+++ demo/docs/service-worker.js
@@ -1,0 +1,2 @@
+if(!self.define){let e,s={};const o=(o,n)=>(o=new URL(o+".js",n).href,s[o]||new Promise((s=>{if("document"in self){const e=document.createElement("script");e.src=o,e.onload=s,document.head.appendChild(e)}else e=o,importScripts(o),s()})).then((()=>{let e=s[o];if(!e)throw new Error(`Module ${o} didn’t register its module`);return e})));self.define=(n,r)=>{const i=e||("document"in self?document.currentScript.src:"")||location.href;if(s[i])return;let l={};const t=e=>o(e,i),u={module:{uri:i},exports:l,require:t};s[i]=Promise.all(n.map((e=>u[e]||t(e)))).then((e=>(r(...e),l)))}}define(["./workbox-5b385ed2"],(function(e){"use strict";e.setCacheNameDetails({prefix:"demo"}),self.addEventListener("message",(e=>{e.data&&"SKIP_WAITING"===e.data.type&&self.skipWaiting()})),e.precacheAndRoute([{url:"css/app.1ed23334.css",revision:null},{url:"css/chunk-vendors.790de279.css",revision:null},{url:"fonts/bootstrap-icons.49c9b852.woff",revision:null},{url:"fonts/bootstrap-icons.f41d849d.woff2",revision:null},{url:"index.html",revision:"b76d31741a457dfc91725debb28624ce"},{url:"js/about.d8398a3f.js",revision:null},{url:"js/app.8eaa8216.js",revision:null},{url:"js/chunk-vendors.7c0c58d8.js",revision:null},{url:"manifest.json",revision:"db7a8a414a8d292161201a6b7865b53e"},{url:"robots.txt",revision:"b6216d61c03e6ce0c9aea6ca7808f7ca"}],{})}));
+//# sourceMappingURL=service-worker.js.map
