+++ demo/src/components/molecules/MDetails/index.ts
@@ -1,0 +1,3 @@
+import MDetailsVue from "./MDetails.vue";
+
+export default MDetailsVue;
