+++ demo/vue.config.js
@@ -2,6 +2,12 @@ const { defineConfig } = require("@vue/cli-service");
 module.exports = defineConfig({
   transpileDependencies: true,
 
+  // github pages
+  outputDir: "docs",
+  assetsDir: "./",
+  publicPath: "./",
+
+  // webpack
   configureWebpack: {
     devServer: {
       historyApiFallback: {
